requirements.txt
absl-py==0.7.0
astor==0.7.1
certifi==2018.11.29
chardet==3.0.4
dlib==19.17.0
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
idna==2.8
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.0.1
numpy==1.16.1
opencv-python==4.0.0.21
Pillow==5.4.1
pkg-resources==0.0.0
protobuf==3.6.1
pyserial==3.4
PySimpleGUI==3.26.0
PyYAML==3.13
requests==2.21.0
scipy==1.2.1
six==1.12.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
tqdm==4.31.1
urllib3==1.24.1
Werkzeug==0.14.1
